Coments endpoints are moved to api/comments_api_views.py
